put guard and logging code on admin delete document

+++ apps/admin/controllers/documents/delete.rb
@@ -3,6 +3,11 @@ module Admin::Controllers::Documents
     include Admin::Action
 
     def call(params)
+      Analytics.record_access(current_user(session), params['id'], '/delete_document/:id', 'app admin')
+      if !current_user(session).admin
+        Analytics.record_access(current_user(session), params['id'], '/delete_document/:id', 'app admin, unauthorized access')
+        redirect_to '/error/0?An error occurred'
+      end
       id = params['id']
       puts "Admin, Delete, id = #{id}".red
       DocumentRepository.destroy_tree(id, [:verbose, :kill])
